tests/wpt/web-platform-tests/encoding-detection/el-windows-1253.tentative.html

<!doctype html>
<title>el windows-1253</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<p>Πρόκειται για δοκιμή κωδικοποίησης χαρακτήρων: Άρης
<!-- I needed to work capital alpha with tonos into the test somehow... --></p>
<script>
setup({explicit_done:true});
onload = function() {
    test(function() {
        assert_equals(document.characterSet, "windows-1253", 'Expected windows-1253');
    }, "Check detection result");
      done();
};
</script>
