third_party/blink/web_tests/external/wpt/encoding-detection/el-windows-1253.tentative-expected.txt

This is a testharness.js-based test.
FAIL Check detection result assert_equals: Expected windows-1253 expected "windows-1253" but got "ISO-8859-7"
Harness: the test ran to completion.
